+++ src/main/java/com/example/k2k_project/Repository/GradeRepository.java
@@ -3,5 +3,5 @@ package com.example.k2k_project.Repository;
 import com.example.k2k_project.Entity.Grade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GradeRepository extends JpaRepository<Grade, Integer> {
+public interface GradeRepository extends JpaRepository<Grade, Long> {
 }
