+++ src/main/java/com/example/k2k_project/Repository/LodgmentRepository.java
@@ -3,5 +3,5 @@ package com.example.k2k_project.Repository;
 import com.example.k2k_project.Entity.Lodgment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LodgmentRepository extends JpaRepository<Lodgment, Integer> {
+public interface LodgmentRepository extends JpaRepository<Lodgment, Long> {
 }
